4850: Main upgrades for partitions splitter to aws v2
