+++ java/partitions/splitterv2/src/main/java/sleeper/splitterv2/core/split/FindPartitionSplitPoint.java
@@ -18,18 +18,18 @@ package sleeper.splitterv2.core.split;
 import com.facebook.collections.ByteArray;
 import org.apache.datasketches.quantiles.ItemsSketch;
 import org.apache.datasketches.quantiles.ItemsUnion;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.Path;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.transfer.s3.S3TransferManager;
 
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.ByteArrayType;
 import sleeper.core.schema.type.PrimitiveType;
-import sleeper.sketches.Sketches;
-import sleeper.sketches.s3.SketchesSerDeToS3;
+import sleeper.sketchesv2.Sketches;
+import sleeper.sketchesv2.store.S3SketchesStore;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
@@ -110,13 +110,12 @@ public class FindPartitionSplitPoint {
         Sketches load(String filename) throws IOException;
     }
 
-    public static SketchesLoader loadSketchesFromFile(Schema schema, Configuration conf) {
-        return (filename) -> new SketchesSerDeToS3(schema).loadFromHadoopFS(new Path(filename), conf);
+    public static SketchesLoader loadSketchesFromFile(Schema schema) {
+        return (filename) -> new S3SketchesStore(S3Client.create(), S3TransferManager.create()).loadFileSketches(filename, schema);
     }
 
-    public static SketchesLoader loadSketchesFromFile(TableProperties tableProperties, Configuration conf) {
-        SketchesSerDeToS3 serDe = new SketchesSerDeToS3(tableProperties.getSchema());
-        return (filename) -> serDe.loadFromHadoopFS(new Path(filename), conf);
+    public static SketchesLoader loadSketchesFromFile(TableProperties tableProperties) {
+        return (filename) -> new S3SketchesStore(S3Client.create(), S3TransferManager.create()).loadFileSketches(filename, tableProperties.getSchema());
     }
 
 }
